requirements.txt
confluent-kafka
pymongo
pandas
python-dotenv
asyncio
json-logging